+++ Solution_finder/capture.h
@@ -45,7 +45,7 @@ public:
 	}
 
 	//call to get new image
-	Mat Capture(float WidthScale=1, float HeightScale=1) {
+	Mat Capture(double WidthScale=1, double HeightScale=1) {
 		// Now update the pixels using BitBlt
 		BitBlt(hdcMem, 0, 0, size.width, size.height, hdcSys, size.x, size.y, SRCCOPY);
 		StretchBlt(hdcMem, 0, 0, WidthScale, HeightScale, hdcMem, 0, 0, size.width, size.height, SRCCOPY); //change SRCCOPY to NOTSRCCOPY for wacky colors !
@@ -53,7 +53,7 @@ public:
 		return matBitmap;
 	}
 
-	void SelectHWND(float scale) {
+	void SelectHWND(double scale) {
 
 
 	}
